Remove override of "userns" when privileged
The fact that `--privileged` requires `--userns=host` when using
Docker's userns-remap feature is an implementation detail. This
might change in future Docker versions and it already isn't the
case for Podman (which supports various values for `--userns` even
when using `--privileged`).

Thus, we shouldn't try to be smarter than the user here and override
their choice of this flag value.